Adding rooms and hotels rest api

+++ src/main/java/com/hostels/model/User.java
@@ -19,7 +19,7 @@ import java.util.List;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String name;
     private String email;
     private String password;
@@ -33,7 +33,7 @@ public class User {
     )
     private final List<Room> bookedRooms = new ArrayList<>();
 
-    public User(int id, String name, String email, String password, Role role) {
+    public User(Long id, String name, String email, String password, Role role) {
         this.id = id;
         this.name = name;
         this.email = email;
@@ -44,11 +44,11 @@ public class User {
     public User() {}
 
     // Getters and Setters
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
